Feature: Add support for bazel credential helper. (#1)
This adds support for the bazel credential helper spec and splits the build into two separate release artifacts that are built on every tag creation.

Now there is a docker-credential-jfrog binary with the docker credential helper, and a bazel-credential-jfrog with the bazel credential helper. The README.md is updated to include both instructions.